feat: hide window when doing actions
